+++ extensions/iterm/src/commands/index.ts
@@ -1,0 +1,6 @@
+export * from "./EditInItermCommand";
+export * from "./NewItermTabCommand";
+export * from "./NewItermWindowCommand";
+export * from "./OpenItermHereCommand";
+export * from "./SplitItermHorizontallyCommand";
+export * from "./SplitItermVerticallyCommand";