+++ extensions/iterm/src/edit-in-iterm-profile.tsx
@@ -1,0 +1,4 @@
+import { EditInItermCommand } from "./commands";
+import { wrapActionWithProfileCommand } from "./utils";
+
+export default wrapActionWithProfileCommand(EditInItermCommand);