+++ extensions/iterm/src/hooks/useMount.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export const useMount = () => {
+  const [mountState, setMountState] = React.useState(false);
+
+  React.useEffect(() => {
+    setMountState(true);
+
+    return () => setMountState(false);
+  }, []);
+
+  return mountState;
+};