+++ extensions/iterm/src/new-iterm-tab-profile.tsx
@@ -1,0 +1,4 @@
+import { NewItermTabCommand } from "./commands";
+import { wrapActionWithProfileCommand } from "./utils";
+
+export default wrapActionWithProfileCommand(NewItermTabCommand);