+++ extensions/iterm/src/split-iterm-horizontally-profile.tsx
@@ -1,0 +1,4 @@
+import { SplitItermHorizontallyCommand } from "./commands";
+import { wrapActionWithProfileCommand } from "./utils";
+
+export default wrapActionWithProfileCommand(SplitItermHorizontallyCommand);