+++ extensions/iterm/src/split-iterm-vertically-profile.tsx
@@ -1,0 +1,4 @@
+import { SplitItermVerticallyCommand } from "./commands";
+import { wrapActionWithProfileCommand } from "./utils";
+
+export default wrapActionWithProfileCommand(SplitItermVerticallyCommand);