+++ extensions/iterm/CHANGELOG.md
@@ -1,5 +1,10 @@
 # iTerm Changelog
 
+## [Extension improvement] - 2024-03-03
+- Total rewrite
+- Added support for Python API
+- Added commands: `New iTerm Window (Profile)`, `New iTerm Tab (Profile)`, `Split iTerm Horizontally (Profile)`, `Split iTerm Vertically (Profile)`, `Edit in iTerm (Profile)`, `Open iTerm Here (Profile)`
+
 ## [Extension improvement] - 2023-10-03
 - Added `Edit in iTerm` and `Open iTerm here` commands